feat: Mejoras en la interfaz y ajustes en la base de datos
Se han realizado cambios en el diseño de las clases `FormLogin` y `FormRecuperarContra`, aumentando el tamaño de varios componentes y ajustando sus posiciones para mejorar la disposición visual de la interfaz de usuario.

Se cambió el tipo de la propiedad `Cant` en `HRemitoDetalleProduccion` de `int` a `decimal` y se actualizaron las consultas SQL en `HRemitoDetalleProduccionRepository` para reflejar el cambio de nombre de la tabla.

En `ucConsultaActividad`, se añadieron nuevos controles para filtros y se implementaron métodos para aplicar y limpiar filtros, mejorando la experiencia del usuario.

Se ajustaron las dimensiones de varios paneles en `FormPrincipal` y se implementaron validaciones en los campos de entrada de `UCConsultaHerramienta` y `UCIngresoHerramienta`.

Finalmente, se actualizaron las inserciones en el archivo SQL de carga de datos para poblar las tablas de herramientas y actividades, así como eliminaciones de registros antiguos.

+++ Agraria.Modelo/Entidades/HRemitoDetalleProduccion.cs
@@ -13,7 +13,7 @@ namespace Agraria.Modelo.Entidades
         public string? Art_Cod { get; set; }
         public string? Descr { get; set; }
         public decimal P_Unit { get; set; }
-        public int Cant { get; set; }
+        public decimal Cant { get; set; }
         public decimal P_X_Cant { get; set; }
 
         public HRemitoDetalleProduccion() { }
